% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBinaryTreeMap.R
\name{createBinaryTreeMap}
\alias{createBinaryTreeMap}
\title{Given a BinaryTree, traverse the nodes and map out the conditional logic}
\usage{
createBinaryTreeMap(t, df = NULL, path = NULL)
}
\arguments{
\item{t}{an object of class BinaryTree}

\item{df}{a data.frame}

\item{path}{a string representing the nodes traversed}
}
\value{
a data.frame representation of a BinaryTree
}
\description{
Given a BinaryTree, traverse the nodes and map out the conditional logic
}
