% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateBinaryTreeMap.R
\name{evaluateBinaryTreeMap}
\alias{evaluateBinaryTreeMap}
\title{Gets the response for a single set of inputs by traversing the tree map}
\usage{
evaluateBinaryTreeMap(treeMap, inputData, node = 1)
}
\arguments{
\item{treeMap}{a minified BinaryTree exported as a data.frame}

\item{inputData}{a data.frame row containing the inputs to be evaluated}

\item{node}{an integer representing the nodeID currently under evaluation}
}
\value{
the repsonse for the given input
}
\description{
Gets the response for a single set of inputs by traversing the tree map
}
