% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateBinaryTreeMap.R
\name{getBinaryTreeResponse}
\alias{getBinaryTreeResponse}
\title{Given a binary tree map, and a data frame of inputs determine the responses}
\usage{
getBinaryTreeResponse(treeMap, inputData)
}
\arguments{
\item{treeMap}{a minified BinaryTree exported as a data.frame}

\item{inputData}{a data.frame containing the inputs to be evaluated}
}
\value{
the repsonses for the given inputs
}
\description{
Given a binary tree map, and a data frame of inputs determine the responses
}
