% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateBinaryTreeMap.R
\name{validateBinaryTreeMap}
\alias{validateBinaryTreeMap}
\title{create a confitional inference tree, minify it and cross-check the responses.
analysis taken from https://datawookie.netlify.com/blog/2013/05/package-party-conditional-inference-trees/}
\usage{
validateBinaryTreeMap()
}
\description{
create a confitional inference tree, minify it and cross-check the responses.
analysis taken from https://datawookie.netlify.com/blog/2013/05/package-party-conditional-inference-trees/
}
