% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteTreeMapCondition.R
\name{writeTreeMapCondition}
\alias{writeTreeMapCondition}
\title{Auto generate the conditional logic from BinaryTree exported as a data.frame by \code{\link{createBinaryTreeMap}}}
\usage{
writeTreeMapCondition(treeMap, node = 1)
}
\arguments{
\item{treeMap}{the exported tree map (data.frame)}

\item{node}{id for recursive evaluation}
}
\value{
string containing the response function
}
\description{
Auto generate the conditional logic from BinaryTree exported as a data.frame by \code{\link{createBinaryTreeMap}}
}
